requirements.txt
pytest>=7.4.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest-html>=3.2.0
pytest-cov>=4.1.0
